Vlad feedback

+++ test/ttmlir/Transforms/reenable_dps.mlir
@@ -4,7 +4,7 @@
 // and a new empty tensor is created instead
 // CHECK-LABEL: func.func @test_broken_dps_simple
 // CHECK-SAME: (%[[ARG0:.*]]: tensor<64x128xf32>, %[[ARG1:.*]]: tensor<64x128xf32>)
-func.func @test_broken_dps_simple(%arg0: tensor<64x128xf32>, %arg1: tensor<64x128xf32>) -> tensor<64x128xf32> attributes {ttir.return_to_output_mapping = 1 : i32} {
+func.func @test_broken_dps_simple(%arg0: tensor<64x128xf32>, %arg1: tensor<64x128xf32>) -> tensor<64x128xf32> attributes {ttir.return_to_output_mapping = 1 : i64} {
   %0 = tensor.empty() : tensor<64x128xf32>
   %1 = linalg.abs ins(%arg0 : tensor<64x128xf32>) outs(%0 : tensor<64x128xf32>) -> tensor<64x128xf32>
 
@@ -15,7 +15,7 @@ func.func @test_broken_dps_simple(%arg0: tensor<64x128xf32>, %arg1: tensor<64x12
 
 // CHECK-LABEL: func.func @test_broken_dps_with_reduce_and_expand
 // CHECK-SAME: (%{{.*}}: tensor<64x128xf32>, %[[ARG1:.*]]: tensor<1x64x128xf32>)
-func.func @test_broken_dps_with_reduce_and_expand(%arg0: tensor<64x128xf32>, %arg1: tensor<1x64x128xf32>) -> tensor<1x64x128xf32> attributes {ttir.return_to_output_mapping = 1 : i32} {
+func.func @test_broken_dps_with_reduce_and_expand(%arg0: tensor<64x128xf32>, %arg1: tensor<1x64x128xf32>) -> tensor<1x64x128xf32> attributes {ttir.return_to_output_mapping = 1 : i64} {
   // Pattern with reduce followed by expand_shape
   %cst = arith.constant 0.0 : f32
   %0 = tensor.empty() : tensor<64x128xf32>
@@ -37,7 +37,7 @@ func.func @test_broken_dps_with_reduce_and_expand(%arg0: tensor<64x128xf32>, %ar
 
 // CHECK-LABEL: func.func @test_noop_case
 // CHECK-SAME: (%[[ARG0:.*]]: tensor<64x128xf32>, %[[ARG1:.*]]: tensor<64x128xf32>)
-func.func @test_noop_case(%arg0: tensor<64x128xf32>, %arg1: tensor<64x128xf32>) -> tensor<64x128xf32> attributes {ttir.return_to_output_mapping = 1 : i32} {
+func.func @test_noop_case(%arg0: tensor<64x128xf32>, %arg1: tensor<64x128xf32>) -> tensor<64x128xf32> attributes {ttir.return_to_output_mapping = 1 : i64} {
   // NOOP case where the return value is directly a block argument
   // CHECK: linalg.copy ins(%[[ARG0]] : tensor<64x128xf32>) outs(%[[ARG1]] : tensor<64x128xf32>)
   return %arg0 : tensor<64x128xf32>
@@ -45,7 +45,7 @@ func.func @test_noop_case(%arg0: tensor<64x128xf32>, %arg1: tensor<64x128xf32>) 
 
 // CHECK-LABEL: func.func @test_already_enabled_dps
 // CHECK-SAME: (%[[ARG0:.*]]: tensor<64x128xf32>, %[[ARG1:.*]]: tensor<64x128xf32>)
-func.func @test_already_enabled_dps(%arg0: tensor<64x128xf32>, %arg1: tensor<64x128xf32>) -> tensor<64x128xf32> attributes {ttir.return_to_output_mapping = 1 : i32} {
+func.func @test_already_enabled_dps(%arg0: tensor<64x128xf32>, %arg1: tensor<64x128xf32>) -> tensor<64x128xf32> attributes {ttir.return_to_output_mapping = 1 : i64} {
   // This function already uses the output buffer, so the pass should skip it
   %0 = linalg.abs ins(%arg0 : tensor<64x128xf32>) outs(%arg1 : tensor<64x128xf32>) -> tensor<64x128xf32>
